fix: allocation bugs + heap visualizer
re-organized test code
heap visualization is now a proper feature instead of code I keep commenting or removing before commits.
No more spreadsheets needed

Since the last tag points to the debug region which is out of bounds as far as the kernel is concerned, it happens to also be where the fake tag is for debugging so heap dump shows that too, accidental feature.

Will change the args so the 3rd to setupProgram is debug region size instead